Implement content iterators and related changes
This is a major feature addition, named 'content iterators'.

libfyaml does not copy token contents, merely keeps track of where
in the input stream those content lie. However up until this commit
in order to process the content one would have to call one of the
*_get_text() methods.

These methods linearize the token content, respecting any direct output
capable tokens (like single line plain scalars) but perform a scan and
copy operation for any others, including block/flow scalars etc.

Which is perfectly fine for many cases, but not for when multi
megabyte (or even gigabyte) yaml files need processing.

The solution is to use token content iterators, which will 'chunk'
the input token without consuming any memory. This in conjuction
with the use of memory mapped files, reduces memory consumption
to a trivial low amount.

As an example, a 160KB yaml file would require about 650KB memory
allocated using standard libyaml (typically libyaml has x4 memory
usage compared to the original file size).
libfyaml will process the same file using only 6.8KB of allocated
memory, which is 100 times less.

Typically memory consumed by libfyaml using the iterator API
is only proportional to the number of nodes in a document, and not
related to the amount of data the nodes contain.

If an application uses the streaming interface, the amount of memory
required is constant, and kept to a very low figure (a couple of Kbytes).

For this to work however, considerable changes had to be made, in
how parsing works, how the emitter works, how parsing has to keep
track of the size of the token.

User level APIs are also provided which the iterators may be used
by an application.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>