Implement circular reference detection.
When resolving a document there must be not circular reference
nor should there be duplicate mapping keys.

For example:
  ---
  - &a [ *b ]
  - &b [ *a ]
  ...

This YAML document is valid if no resolution has taken place.
However if we try to resolve obviously it cannot be valid.

Similarly:
  ---
  &l a: b
  *l : c
  ...

This document is valid when not resolved (since *l != a) but
when resolved it would turn to, which is invalid.
  ---
  a: b
  a: c
  ...

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>