Implement event based emitter functionality.
According to popular request, this commit implements an
event based emitter method, using fy_emit_event()

It implements a streaming output method, which means that
functionality that depends on the presence of a tree is
not available, i.e. sorting of keys of a mapping, checking that
no duplicate keys exist neither that aliases point to somewhere
valid.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>