Implement a relative simple dejson emit method
Implement a de-JSONify emit method. Convert JSON documents
to YAML, while attempting to convert as many scalars to plain
style, which makes things much more readable.

For example the original JSON

	{ "foo": "bar", "baz": "frooz" }

would keep the formatting even when output is set to block mode:

	"foo": "bar"
	"baz": "frooz"

Using dejson mode

	foo: bar
	baz: frooz

It is clever enough to keep formating as strings JSON strings that
contain numbers/null/booleans, so that a subsequent conversion to json
would work as expected.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>