Major refactoring of diagnostics
Very large refactoring of diagnostics; we introduce an
atom raw line iterator and use it to mark areas of interest
in the input stream. We also provide diagnostics for fy_nodes()
which required the introduction of a new token type FYTT_INPUT_MARKER.

We also export a diagnostic method to users, for when they need to
do validation themselves and mark a node as an output.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>