Introduce pure JSON input mode
Introduce a pure JSON input mode (along with the switches to
enable it). This mode marks the parser input strictly as JSON.

With this commit the JSONTestSuite passes without failures.

Note that by default the json detection is set at auto; which
means that files with a .json extension are parsed as JSON.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>