Add generational markers on input and atoms
When using growable inputs (i.e. using realloc) direct pointers to
token text become (potentially) invalid if realloc changes the buffer
allocated. During normal operation this is seldom done, but it is
apparent when using valgrind.

Add a generational marker to each input which is incremented when
the buffer potentially changes, and check when accessing the token's
text pointer cached.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>