Introduce a reader construct and use it for parsing
We're moving to a system that's going to be able to parse multiple
variants of inputs, which makes the original reader implementation
unsuitable, due to the mingling of reading and parsing of YAML/JSON.

The reader construct allows us to separate YAML from the task of
reading, but this requires the use of readers instead of parsers
in places.

Diagnostic methods for the reader is provided as well.

Care has been taken for the external library interface to not change
at all, so things should continue to work as previously for users.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>