Introduce a YPATH parser interface.
Introduce a YPATH parser interface which allows us to parse
YPATH expressions.

YPATH expressions differ than the original (and still in use)
path reference API in that it's a query/filter interface.

The original path syntax only returns a single node reference,
but the YPATH form is capable of more.

For example the following expression is equivalent.

/foo

Refers to the mapping value with key foo of the map at the root.

While the following is only valid in YPATH

/*

Refers to every child of the root node.

Brief reference

/		- At the start of the expression refers to the root
		- At the end of the expression is the collection filter
		- Anywhere else is the chaining operator
^		- The root
.		- This node
..		- The parent of this node
*		- Every child of this node
**		- Every child of this node recursively
:<key>		- Sibling of this node with <key>
$		- Scalar filter (i.e. remove all non scalar results)
%		- Collection filter
[]		- Sequence filter
{}		- Mapping filter
,		- Or operator (/foo,bar refers to either foo or bar at the root)
"..." or '...'	- Flow scalar key
{...} or [...]	- Flow mapping or sequence key

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>