Foreign language bindings support implemented
As a fact of life, most foreign languages (other than C/C++) already
use bindings based on libyaml. So in order to make things less
painful for authors of bindings, this patch introduces changes that
provide a comparable API.

That included exporting a lot of internal details, which frankly
it's better to not be used. Care has been given in not to over-export
internal details, and when data are returned, it is done via
accessor methods.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>