Removal of diagnostic option from parser config flags
This major patch completely removes all FYPCF_DEBUG and FYPCF_COLOR
options from the parser config flags. It was arguably a bad decision
to put them there, since they rightfully belong to the diagnostic
object, as they are used there.

Care has been taken to define the old options as zeros so that
code can still continue to compile (although those options are NOPed).
Eventually it is expected that by the time of 1.0 all these remnants
will be removed.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>