Fully implement YAML version handling and new API
Until now there was no YAML API version, the strategy was to parse
anything, and error out according to YAML 1.2 semantics.

This patch adds a version API that applications can use in order
to finely control the parsing according to version quirks.

There are not many quirks now but this is the way to go forward
in the future where quirks will be extensive.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>