Complete support for YAML1.1 parsing
This large patchset adds support for YAML1.1.
It is quite large and intrusive because the 'content' of what
a linebreak is must propagate all around in order to be handled
correctly.

This makes libfyaml parse (in YAML1.1 mode) all the examples of
the 1.1 spec.

Note, just because unicode linebreaks are handled now correctly,
it is absolute bonkers to try to use them in any other context than
in block literals.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>