Add a internal for now non recursive iterator.
Recursive iteration is fine, but does have depth limitations due
to the amount of memory consumer by a stack frame.

Introduce an internal node iterator that should handle arbitrary
amount of depth, and be significantly more memory efficient.

Eventually this non-recursive iterator will be the default, but
for now leave both of them.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>