Major update of YPATH support
This arguably large patch completely overhauls the implementation
of YPATH in libfyaml while simultaneously exporting the capability
for use in external users.

This functionality is controlled by the use of two flag bits,
the FYNWF_PTR_YPATH bit in fy_node_walk_flags and the
FYPCF_YPATH_ALIASES bit in the parser configuration.

The first enables the use of YPATH expressions as paths in the
methods that use fy_node_walk_flags, while the latter interpets the
aliases present in a parsed document as YPATH expressions.

For now alias resolution only resolves to the first of the resulting
nodes of a path expression execution, but that may change in the
future.

By default all this functionality is off.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>