Major internal composer and path reorg
With this patch, the composer is now ready for exporting
a limited interface based on a single callback on the parser.

Using fy_parse_compose() it is possible to build a document
using a simplified callback and path query that makes it
considerably easier than raw event processing.

In a nutshell, the path encapsulates the path and the parsing
state in a way that is natural for bindings to use.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>