Implement a publically visible document iterator
A document iterator is a non-recursive method to traverse a document.

You can either traverse a document generating events from it, or
just iterate over it's nodes.

This is a demonstration of the functional equivalence of a finite
event stream and document tree containing nodes created from that
stream. The generated event stream is identical to the one that created
the document and no information is lost.

This is helpful for transformation from one to the other form, according
to application requirements, without losing information.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>