Add --document-event-stream option to the testsuite tool command
The --document-event-stream option for the testsuite tool command
changes the way that the testsuite events are generated.

Instead of directly emitting them as generated by the parser, a
regular document is created and then this is used as the source
for a document iterator generating events and driving the testsuite
output.

There should be no functional change when the input is correct, and
no resolution is performed.

If resolution is performed then the test suite events are the ones that
would be generated for the resulted resolved document.

While this is marginally useful for the testsuite tool, it is of great
importance as a demonstrator of a technique of using the library for
performing resolution and generating an event stream out of the resulted
document. This is the regular case when using libfyaml in a bindings to
non C based languages, where this means that there is no need for the
binding to handle anchor and alias at all.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>