Make composer respect the resolve document flag.
Now that the document iterator infrastructure is in place we can
finally implement automatic resolver functionality using the
composer interface.

If the FYPCF_RESOLVE_DOCUMENT flag is not set, then operation of
the composer is unchanged.

If the flag is set however, parsing is done via producing a document,
which is subsequently resolved, and the result is use to generate
events for the composer.

In essense consumers never need to handle resolution, and deal with
an already resolved event stream.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>