Correctly handle block literals ended with EOF
Things get complicated when the stream ends without an LB.

Block literals are especially hairy, and this patch handle those
corner cases. Note that it's really hard to generate a non LB
terminated file in most editors, but nevertheless it might happen.

This fixes the JEF9/02 testcase in the upcoming testsuite.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>