emit-pretty: Introduce emit-pretty
Pretty mode tries to output prettier YAML (most of the times).
It is similar to dejson and has tries to do the following.

For scalars that contain linebreaks but otherwise are 'plain' use
a block literal style.

i.e. for --- "foo\nbar\n" it will output:
--- |
  foo
  bar

Previously it used a plain style which is confusing
---
  foo

  bar

And yes, the contents of the scalar in both cases are identical.

Also it will output a document start indicator at the start of
a scalar only document.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>