input: No more infinite buffers for streaming mode inputs
For streaming mode inputs, where input may be unbounded we created
infinite buffers. In these cases a streaming mode is used and not
a document, so this was suboptimal to say the least.

This patch enables a choping mode set at 16 times the page size.
If the input passes the choping mark (at a token boundary) a new
input will be allocated and the old input will be unrefed.

Note that tokens take references to the input, so while references
exist the memory will still be around.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>