Do not attempt complex keys for JSON
JSON files have a tendency to be large and in one line.
YAML has complex keys i.e. { foo: bar }: baz.

The method to detect this is to wait until the enclosing
flow marker, or a line change, which in those kind of cases
causes the whole file to be read (and tokens buffered) before
emitting the event.

This patch disables this scan ahead for json which is the right
thing to do.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>