Add composer halt feature.
When we stop the composer, we didn't had a way for users to free up
their data. We do it now via the pumping of FYET_NONE messages.

Note that this is a slight API change, and clients may break.
If that's the case; just ignore the FYET_NONE event in your handlers.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>