parse: Complete implementation of merge key in streaming mode
It is quite complicated but now merge keys work as the 1.1 spec
states. Note that YAML 1.2 doesn't support merge keys which is
the default. To use them either the YAML file must state that
it's version 1.1 or forcing 1.1.

I'm still not sold on the idea of merge keys, but now the 1.1
support is complete for streaming mode.

Streaming mode is at least x10 faster and consumes basically zero
memory, so this is a huge win for end users.

Expect the default for fy-tool to switch to streaming mode and
the original document mode to be used only as an API for dynamic
YAML creation.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>