Add a work stealing push/pull threading pool implementation
To make parallel tasks done we need a threading pool library.
Unfortunately haven't found someone that would cut it, especially
when the goal is to be competitive with rust's Rayon implementation.

So this commit introduces a very fast threading pool that implements
Rayon join semantics (but with joins being arbitrarily sized, not
just 2).

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>