Allow prefix only tags to pass `fy_tag_scan`.
Allows tags like `!` to pass scanning.
This is required to pass the emitter tests when the
events are passed through the emit event apis.

In `fy_tag_scan` it replaces the test for `handle_length <= 0`,
with just less than zero.

This change is validated by the fact that further down in
`fy_tag_scan` there is a check for `handle_length == 0’,
meaning it is expecting zero-length handles and handles them properly.

Signed-off-by: Kevin Wooten <kevin@wooten.com>