Force doc end marker when “pretty” emit mode forces a doc start marker
When emitting with pretty mode enabled, and a single scalar root object
is output, it forces a document start marker to be output.

This change tracks when a doc start marker is forced and forces a doc end
marker to match.

This change only affects pretty mode because this is the only time when a
doc start marked is forced out (even when it is requested to be implicit).

Signed-off-by: Kevin Wooten <kevin@wooten.com>