Add logical not operator in walk operations
Remove from input using the ! operator at the start of the expression.
Things are very messy, and syntax is in flux, but at least

  !/foo/bar

Will remove /foo/bar from the input

  !/foo/bar,baz

Will remove /foo/bar and foo/baz from the input

  !(/foo),(/bar/baz)

Will remove /foo and /bar/baz from the input.

This is highly experimental for now.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>