allocators: Introduce allocator infrastructure
The allocator infrastructure is intented to be used for
fast and optionally deduplicated object storage.

The current infra supports the following allocators

- linear: simple linear allocator storing objects in sequence
- malloc: Uses malloc to store each object. Inefficient but
  interfaces best with ASAN or valgrind in order to find bugs.
- mremap: Uses mmaped/mremaped arenas. Best for performance.

Each of the above can be paired with a dedup allocator which
stores objects only once.

Finally
- auto: An auto allocator selecting what best for a given scenario.

Note that these are tuned for handling frees on a tag granularity
and not per object. If you need per object freeing you're better
off with the standard malloc() allocator.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>